Add initial access denied page and move account maintenance values to their own component
